+++ packages/react-universal/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Next
+## 2.0.0-beta.7
 
 * `useNavigationHandler` & `useStatefulNavigationHandler` re-use event handler callback when possible. They also take a `target` string instead of the full `forward` props object.